+++ fs/rsext4/src/blockgroup_description/mod.rs
@@ -2,7 +2,7 @@
 
 mod desc;
 mod disk;
-#[cfg(axtest)]
+#[cfg(all(axtest, feature = "axtest"))]
 pub(crate) use self::disk::block_group_desc_disk_format_rules_hold_for_test;
 mod stats;
 mod table;